+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Declare the originating types for generated `*Builder` class to make it easier for tooling to determine when to re-run annotation processor to generate builder class.
+
 ### [v0.139](https://github.com/react4j/react4j/tree/v0.139) (2019-11-05) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.138...v0.139)
 
 Changes in this release: